Refactor: 영양성분 변수 타입 변경 (#8)
float에서 double로 변경

+++ src/main/java/com/petplate/petplate/common/EmbeddedType/Vitamin.java
@@ -14,19 +14,19 @@ import lombok.NoArgsConstructor;
 public class Vitamin {
 
     @Column(nullable = false)
-    private float vitaminA;
+    private double vitaminA;
 
     @Column(nullable = false)
-    private float vitaminB;
+    private double vitaminB;
 
     @Column(nullable = false)
-    private float vitaminD;
+    private double vitaminD;
 
     @Column(nullable = false)
-    private float vitaminE;
+    private double vitaminE;
 
     @Builder
-    public Vitamin(float vitaminA, float vitaminB, float vitaminD, float vitaminE) {
+    public Vitamin(double vitaminA, double vitaminB, double vitaminD, double vitaminE) {
         this.vitaminA = vitaminA;
         this.vitaminB = vitaminB;
         this.vitaminD = vitaminD;
